refactor(Person, PersonName): PersonName 분리 작성
